Python - Smart Librarian Project/requirements.txt
openai>=1.30.0
chromadb>=0.5.0
python-dotenv>=1.0.1
streamlit>=1.35.0
pytest